requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
streamlit
matplotlib
seaborn
plotly
google-auth
wordcloud
folium
streamlit-folium
